# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Interlock API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/approval-requests/{request_id}": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "Get Review",
        "operationId": "get_review_approval_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/approval-requests/{request_id}/open": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Open Review",
        "operationId": "open_review_approval_requests__request_id__open_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "x-actor-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "x-actor-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/approval-requests/{request_id}/cancel": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Cancel Review",
        "operationId": "cancel_review_approval_requests__request_id__cancel_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "x-actor-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "x-actor-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/approval-requests/{request_id}/preview": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Preview Review",
        "description": "Render without approving -- a pure function of current data, safe to\ncall repeatedly as the user toggles recipients before committing.",
        "operationId": "preview_review_approval_requests__request_id__preview_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/approval-requests/{request_id}/commit": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Commit Review",
        "operationId": "commit_review_approval_requests__request_id__commit_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "x-actor-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "x-actor-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalKind": {
        "type": "string",
        "enum": [
          "MORNING",
          "EVENING",
          "MANUAL"
        ],
        "title": "ApprovalKind"
      },
      "ApprovalRequestOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_id": {
            "type": "string",
            "title": "Display Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ApprovalKind"
          },
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "scheduled_for": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled For"
          },
          "state": {
            "$ref": "#/components/schemas/ApprovalState"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "approved_by_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By User Id"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_id",
          "kind",
          "local_date",
          "scheduled_for",
          "state",
          "opened_at",
          "approved_by_user_id",
          "approved_at"
        ],
        "title": "ApprovalRequestOut"
      },
      "ApprovalState": {
        "type": "string",
        "enum": [
          "DRAFT",
          "REVIEW_PENDING",
          "USER_EDITING",
          "READY",
          "APPROVED",
          "CLOSED_NO_SHARE",
          "CANCELLED",
          "EXPIRED"
        ],
        "title": "ApprovalState",
        "description": "One scheduled review. Ends when the user decides what to do with it."
      },
...
      "ChangeSummaryOut": {
        "properties": {
          "added": {
            "type": "integer",
            "title": "Added"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "modified": {
            "type": "integer",
            "title": "Modified"
          },
          "overdue": {
            "type": "integer",
            "title": "Overdue"
          },
          "still_pending": {
            "type": "integer",
            "title": "Still Pending"
          }
        },
        "type": "object",
        "required": [
          "added",
          "completed",
          "modified",
          "overdue",
          "still_pending"
        ],
        "title": "ChangeSummaryOut"
      },
      "CommitMode": {
        "type": "string",
        "enum": [
          "UPDATE_ONLY",
          "UPDATE_AND_SHARE",
          "SHARE_ONLY"
        ],
        "title": "CommitMode"
      },
      "CommitRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/CommitMode"
          },
          "action_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Action Version"
          },
          "task_updates": {
            "items": {
              "$ref": "#/components/schemas/CommitTaskUpdate"
            },
            "type": "array",
            "title": "Task Updates"
          },
          "recipient_group_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recipient Group Ids"
          },
          "send_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Send At"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "template_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "action_version"
        ],
        "title": "CommitRequest"
      },
      "CommitResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "mode": {
            "$ref": "#/components/schemas/CommitMode"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareJobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/ShareRecipientOut"
            },
            "type": "array",
            "title": "Recipients"
          },
          "preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview"
          },
          "already_committed": {
            "type": "boolean",
            "title": "Already Committed"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "mode",
          "job",
          "recipients",
          "preview",
          "already_committed"
        ],
        "title": "CommitResponse"
      },
      "CommitTaskUpdate": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "changes": {
            "$ref": "#/components/schemas/TaskEditFields"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "changes",
          "expected_version"
        ],
        "title": "CommitTaskUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PreviewResponse": {
        "properties": {
          "rendered_body": {
            "type": "string",
            "title": "Rendered Body"
          }
        },
        "type": "object",
        "required": [
          "rendered_body"
        ],
        "title": "PreviewResponse"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "URGENT"
        ],
        "title": "Priority"
      },
      "RecipientState": {
        "type": "string",
        "enum": [
          "QUEUED",
          "SENDING",
          "SENT",
          "FAILED",
          "SKIPPED",
          "CANCELLED"
        ],
        "title": "RecipientState",
        "description": "One group, one job. The actual unit of delivery."
      },
...
      "ReviewDetailOut": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/ApprovalRequestOut"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "type": "array",
            "title": "Tasks"
          },
          "summary": {
            "$ref": "#/components/schemas/TaskSummaryOut"
          },
          "changes_since_last_share": {
            "$ref": "#/components/schemas/ChangeSummaryOut"
          }
        },
        "type": "object",
        "required": [
          "request",
          "tasks",
          "summary",
          "changes_since_last_share"
        ],
        "title": "ReviewDetailOut"
      },
      "ShareJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_id": {
            "type": "string",
            "title": "Display Id"
          },
          "state": {
            "$ref": "#/components/schemas/ShareJobState"
          },
          "action_version": {
            "type": "integer",
            "title": "Action Version"
          },
          "deferred_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deferred Reason"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_id",
          "state",
          "action_version",
          "deferred_reason",
          "sent_at"
        ],
        "title": "ShareJobOut"
      },
      "ShareJobState": {
        "type": "string",
        "enum": [
          "PENDING",
          "SCHEDULED",
          "RESCHEDULED",
          "DEFERRED",
          "SENDING",
          "SENT",
          "PARTIALLY_SENT",
          "FAILED",
          "CANCELLED",
          "SUPERSEDED"
        ],
        "title": "ShareJobState",
        "description": "One approved send. Fans out to one or more recipients."
      },
      "ShareRecipientOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "whatsapp_group_id": {
            "type": "string",
            "title": "Whatsapp Group Id"
          },
          "state": {
            "$ref": "#/components/schemas/RecipientState"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "client_message_id": {
            "type": "string",
            "title": "Client Message Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "type": "string",
            "title": "Error Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "whatsapp_group_id",
          "state",
          "attempts",
          "client_message_id",
          "error_code",
          "error_detail"
        ],
        "title": "ShareRecipientOut"
      },
...
      "TaskEditFields": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TaskEditFields",
        "description": "Every field PATCH /tasks/{id} may change. All optional -- only fields\nactually present in the request body are applied, via\n``model_dump(exclude_unset=True)``, so \"not sent\" and \"explicitly set to\nits current value\" are properly distinguishable from \"clear this field\"."
      },
      "TaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_id": {
            "type": "string",
            "title": "Display Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "remarks": {
            "type": "string",
            "title": "Remarks"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "$ref": "#/components/schemas/TaskSourceKind"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "last_shared_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Shared At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_id",
          "title",
          "description",
          "project_id",
          "owner_id",
          "priority",
          "status",
          "created_at",
          "updated_at",
          "due_date",
          "completed_at",
          "remarks",
          "tags",
          "source",
          "version",
          "last_shared_at"
        ],
        "title": "TaskOut"
      },
      "TaskSourceKind": {
        "type": "string",
        "enum": [
          "APP",
          "GOOGLE_SHEETS",
          "EXCEL",
          "IMPORT"
        ],
        "title": "TaskSourceKind",
        "description": "Where a task last came from. Used for audit, not for routing."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "IN_PROGRESS",
          "COMPLETED",
          "BLOCKED",
          "DEFERRED"
        ],
        "title": "TaskStatus"
      },
      "TaskSummaryOut": {
        "properties": {
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "in_progress": {
            "type": "integer",
            "title": "In Progress"
          },
          "blocked": {
            "type": "integer",
            "title": "Blocked"
          },
          "deferred": {
            "type": "integer",
            "title": "Deferred"
          },
          "overdue": {
            "type": "integer",
            "title": "Overdue"
          },
          "due_today": {
            "type": "integer",
            "title": "Due Today"
          },
          "newly_added": {
            "type": "integer",
            "title": "Newly Added"
          },
          "modified_today": {
            "type": "integer",
            "title": "Modified Today"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "completed",
          "pending",
          "in_progress",
          "blocked",
          "deferred",
          "overdue",
          "due_today",
          "newly_added",
          "modified_today",
          "total",
          "remaining"
        ],
        "title": "TaskSummaryOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}